Source/requirements.txt
certifi 
colorama 
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
fvcore==0.1.5.post20210915
imageio==2.22.1
iopath==0.1.9
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.0
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
networkx==2.8.7
numpy
matplotlib
tqdm
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
plotly==5.10.0
portalocker
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
PyWavelets==1.4.1
PyYAML
scikit-learn==1.1.2
scipy==1.9.1
six 
sklearn==0.0
tabulate
tenacity==8.1.0
termcolor
threadpoolctl==3.1.0
tifffile==2022.8.12
torch==1.7.1
torchvision==0.8.2
tqdm 
typing_extensions 
yacs 

pytorch3d
scikit-image==0.16.2

